+++ examples/get_list_of_status_updates.php
@@ -12,13 +12,25 @@ use SammyK\FacebookQueryBuilder\FacebookQueryBuilderException;
 try
 {
     $statuses = $fqb->object('me/statuses')->limit(10)->get();
-
-    $statuses->each(function ($v)
-    {
-        var_dump($v['message']);
-    });
 }
 catch (FacebookQueryBuilderException $e)
 {
+    echo '<p>Error: ' . $e->getMessage() . "\n\n";
+    echo '<p>Facebook SDK Said: ' . $e->getPrevious()->getMessage() . "\n\n";
+    echo '<p>Graph Said: ' .  "\n\n";
     var_dump($e->getResponse());
+    exit;
+}
+
+if (count($statuses) > 0)
+{
+    echo '<h1>Last 10 Statuses Response:</h1>' . "\n\n";
+    foreach ($statuses as $status)
+    {
+        var_dump($status['message']);
+    }
+}
+else
+{
+    echo 'No statuses returned. Make sure you have the "read_stream" extended permission for this access token.' . "\n\n";
 }