+++ examples/post_a_status_update_then_like_it_then_comment_on_it.php
@@ -14,20 +14,29 @@ try
     // Post the status update
     $status_update = ['message' => 'My witty status update.'];
     $response = $fqb->object('me/feed')->with($status_update)->post();
-    var_dump('Post Status Update Response:', $response);
+    echo '<h1>Post Status Update Response:</h1>' . "\n\n";
+    var_dump($response->toArray());
 
     $status_update_id = $response['id'];
 
     // Like it!
     $response = $fqb->object($status_update_id . '/likes')->post();
-    var_dump('Like Status Update Response:', $response);
+    echo '<h1>Like Status Update Response:</h1>' . "\n\n";
+    var_dump($response->toArray());
 
     // Comment on it
     $comment = ['message' => 'My witty comment on your status update.'];
     $response = $fqb->object($status_update_id . '/comments')->with($comment)->post();
-    var_dump('Post Comment Response:', $response);
+    echo '<h1>Post Comment Response:</h1>' . "\n\n";
+    var_dump($response->toArray());
 }
 catch (FacebookQueryBuilderException $e)
 {
+    echo '<p>Oops! Make sure you have the "publish_actions" extended permission for this access token.' . "\n\n";
+
+    echo '<p>Error: ' . $e->getMessage() . "\n\n";
+    echo '<p>Facebook SDK Said: ' . $e->getPrevious()->getMessage() . "\n\n";
+    echo '<p>Graph Said: ' .  "\n\n";
     var_dump($e->getResponse());
+    exit;
 }