+++ CHANGELOG.md
@@ -1,5 +1,17 @@
-CHANGELOG
-=========
+# CHANGELOG
+
+
+## 1.0.3 - May 30, 2014
+
+- Added feature to obtain an access token via redirect, canvas or Javascript.
+- Added a setter to inject an existing `FacebookSession` object.
+- Better exception handling.
+- Beefed up the examples.
+
+
+## 1.0.2 - May 29, 2014
+
+- Changed SDK dependency to latest stable instead of dev-master.
 
 
 ## 1.0.1 - May 8, 2014